requirements.txt
async-timeout==3.0.1
beautifulsoup4==4.8.0
discord-timers==0.1.0
discord.py==1.3.2
disputils==0.1.3
numpy==1.17.2
oauthlib==3.1.0
pre-commit==1.16.1
public==2019.4.13
pycodestyle==2.5.0
pymongo==3.9.0
python-dotenv==0.10.3
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
urllib3==1.25.6
virtualenv==16.7.5
websockets==8.1
youtube-dl==2020.3.8
zipp==0.6.0
git+https://github.com/Rapptz/discord.py